torch library

+++ csrc/quantization/awq/gemm_kernels.cu
@@ -7,7 +7,7 @@ Shang and Dang, Xingyu and Han, Song}, journal={arXiv}, year={2023}
 }
  */
 
-#include <torch/extension.h>
+#include <torch/all.h>
 #include <c10/cuda/CUDAGuard.h>
 
 #include "dequantize.cuh"
@@ -762,8 +762,8 @@ __global__ void __launch_bounds__(64) group_gemm_forward_4bit_cuda_m16nXk32(
 
 torch::Tensor awq_dequantize(torch::Tensor _kernel,
                              torch::Tensor _scaling_factors,
-                             torch::Tensor _zeros, int split_k_iters, int thx,
-                             int thy) {
+                             torch::Tensor _zeros, int64_t split_k_iters, int64_t thx,
+                             int64_t thy) {
   int in_c = _kernel.dim() == 2 ? _kernel.size(0) : _kernel.size(1);
   int qout_c = _kernel.dim() == 2 ? _kernel.size(1) : _kernel.size(2);
   int num_experts = _kernel.dim() == 2 ? 1 : _kernel.size(0);
@@ -825,7 +825,7 @@ torch::Tensor awq_dequantize(torch::Tensor _kernel,
 
 torch::Tensor awq_gemm(torch::Tensor _in_feats, torch::Tensor _kernel,
                        torch::Tensor _scaling_factors, torch::Tensor _zeros,
-                       int split_k_iters) {
+                       int64_t split_k_iters) {
   int num_in_feats = _in_feats.size(0);
   int num_in_channels = _in_feats.size(1);
   const at::cuda::OptionalCUDAGuard device_guard(device_of(_in_feats));
@@ -888,7 +888,7 @@ torch::Tensor awq_fused_moe(torch::Tensor _in_feats, torch::Tensor _kernel,
                             torch::Tensor _sorted_token_ids_ptr,
                             torch::Tensor _expert_ids_ptr,
                             torch::Tensor _num_tokens_post_padded,
-                            bool mul_weights, int split_k_iters) {
+                            bool mul_weights, int64_t split_k_iters) {
   int num_in_feats = _in_feats.size(0);
   int pad_num_in_feats = _sorted_token_ids_ptr.size(0);
   int num_in_channels = _in_feats.size(2);
